Добавлен запрос выводящей информацию о настройках расположения документов

+++ CQRS/Commands/Handlers/Contenders/RemoveContenderHandler.cs
@@ -9,7 +9,7 @@ using System.Threading.Tasks;
 
 namespace CQRS.Commands.Handlers.Contenders
 {
-    public class RemoveContenderHandler : IRequestHandler<RemoveContenderCommand, int>
+    public class RemoveContenderHandler : IRequestHandler<RemoveContenderCommand, bool>
     {
         private readonly IRepository<Contender> _contenderRepository;
         private readonly IRepository<Option> _optionRepository;
@@ -20,17 +20,26 @@ namespace CQRS.Commands.Handlers.Contenders
             _optionRepository = optionRepository;
         }
 
-        public async Task<int> Handle(RemoveContenderCommand request, CancellationToken cancellationToken)
+        public async Task<bool> Handle(RemoveContenderCommand request, CancellationToken cancellationToken)
         {
             var contender = await _contenderRepository.FindAsync(request.Id);
             if (contender != null)
             {
-                string path = _optionRepository.GetAllAsNoTracking().FirstOrDefault(o => o.PropertyName == OptionTypes.DocumentsSource).Value + "\\" + contender.DocumentSource;
-                File.Delete(path);
-                await _contenderRepository.RemoveAsync(contender);
-                await _contenderRepository.SaveAsync();
+                var documentSource = _optionRepository
+                    .GetAllAsNoTracking()
+                    .FirstOrDefault(
+                    o => o.PropertyName == OptionTypes.DocumentsSource);
+
+                if (documentSource != null)
+                {
+                    string path = documentSource.Value + "\\" + contender.DocumentSource;
+                    File.Delete(path);
+                    await _contenderRepository.RemoveAsync(contender);
+                    await _contenderRepository.SaveAsync();
+                    return true;
+                }
             }
-            return request.Id;
+            return false;
         }
     }
 }
